verify(S-19.03): fix VP-097 Kani harness module path so proofs compile under --cfg kani (F-P3-001)
The VP-097 harness was wired as an out-of-line file-module
(`#[cfg(kani)] mod path_util_kani;` inside path_util.rs). Because
path_util.rs is a non-mod.rs submodule (`pub mod path_util;` in
host/mod.rs), Rust's module-file rule resolves that declaration to
host/path_util/path_util_kani.rs, while the harness file lived at the
sibling host/path_util_kani.rs. Under `--cfg kani` this is E0583
(file not found), so the 4 proofs (H1 no-`..`, H2 containment,
H3 escape-witness, H4 determinism) never compiled or ran.

Fix (OPTION A — convention-align): inline the harness into path_util.rs
as `#[cfg(kani)] mod kani_proofs { ... }`, matching the established
in-crate Kani convention (partition.rs, aggregator.rs kani_proofs;
resolver_classify_trap.rs kani_harnesses — all inline, S-19.03 was the
sole out-of-line file-module). Deleted the separate path_util_kani.rs
and corrected the harness comment, which had falsely claimed to
"mirror the established repo convention" while being the one file that
deviated from it. Proof content unchanged.

Verification (module-resolution only; `cargo kani` itself remains
blocked by the D-826 MSRV gap — Kani-bundled rustc 1.93 < workspace
1.95 — which is a tracked repo-wide item, not this fix):
  cargo rustc -p factory-dispatcher --lib -- --cfg kani
  (applies --cfg kani to the target crate only, not dependencies)
    BEFORE: error[E0583] file not found for module `path_util_kani`
            (help: create host/path_util/path_util_kani.rs)
    AFTER:  E0583 gone; only E0433 "cannot find crate `kani`" remain
            (82, all toolchain-gated — the kani crate is absent under a
            plain --cfg build); kani_proofs now compiles inline in
            path_util.rs (17 kani:: refs cite path_util.rs, was 0).

Gates (harness is cfg(kani)-excluded from normal builds):
  cargo fmt --check --all           PASS
  cargo clippy --workspace --all-targets -- -D warnings   PASS
  cargo test --workspace --all-targets                    PASS
